requirements.txt
fastapi>=0.115.0,<1.0.0
google-adk>=0.5.0
google-genai>=1.0.0
python-dotenv>=1.0.0
uvicorn[standard]>=0.34.0,<1.0.0

